v0.7

+++ src/main/kotlin/PubFightAdventures/Enemy.kt
@@ -1,4 +1,4 @@
 package PubFightAdventures
 
-open class Enemy(var name: String, var hp: Int) {
+open class Enemy(var name: String, var hp: Double) {
 }
